Add client http retrofit and prediction service

+++ site/src/main/java/fr/lernejo/travelsite/controllers/dto/TravelResponse.java
@@ -1,5 +1,5 @@
 package fr.lernejo.travelsite.controllers.dto;
 
 public record TravelResponse(String country,
-                             Integer temperature) {
+                             double temperature) {
 }
